Deactivated the toolbar delegate synchronisation and other toolbar modifications
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@20001 72102866-910b-0410-8b05-ffd578937521

+++ Headers/Additions/GNUstepGUI/GSToolbarView.h
@@ -62,7 +62,7 @@ typedef enum {
   NSToolbarDisplayMode _displayMode;
   NSToolbarSizeMode _sizeMode;
   NSRect _rectAvailable;
-  unsigned int _heightFromLayout;
+  float _heightFromLayout;
 }
 
 - (id) initWithFrame: (NSRect)frame;
